Refactor declaration header grammar boundaries
Separate pure declaration-header classification from the index builder's single-pass traversal, trust propagation, and span finalization.
